+++ R/size_ticks.R
@@ -1,9 +1,9 @@
-#' Compute Spans for Ticks on custom Axes
+#' Compute Spans for Ticks on Custom Axes
 #'
 #' @param breaks Vector of breaks for a ggplot object
 #' @param span  Proportion of the range of the breaks the ticks should span in each direction
 #'
-#' @return Value representing the distance left or right the tick should span from the drawn acis
+#' @return Value representing the distance left or right the tick should span from the drawn axis
 #'
 #'
 #'